# specs/001-proposal-assistant/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: EasyGrant Smart Proposal Assistant API
  description: |
    REST API for the Smart Proposal Assistant demo. Supports funding call upload,
    requirements extraction, RAG-based section generation with citations, sticky edits,
    and DOCX export.
  version: 1.0.0
  contact:
    name: EasyGrant Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://easygrant.onrender.com
    description: Production (Render)
...
paths:
...
  /api/requirements/{session_id}:
    get:
      tags: [Requirements]
      summary: Get extracted requirements checklist
      description: |
        Retrieve the structured requirements blueprint extracted from the
        funding call. Includes sections, word limits, eligibility, scoring.
        (FR-001)
      operationId: getRequirements
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Requirements blueprint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementsBlueprint'
        '404':
          description: Session or funding call not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/sections/{section_id}:
    get:
      tags: [Sections]
      summary: Get section details
      description: Retrieve generated section with current text, citations, and metadata
      operationId: getSection
      parameters:
        - name: section_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Section details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedSectionResponse'
        '404':
          description: Section not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    patch:
      tags: [Sections]
      summary: Update section (user edits)
      description: |
        Apply user edits to a section. Edits are stored separately to enable
        sticky edit preservation during regeneration. (FR-009)
      operationId: updateSection
      parameters:
        - name: section_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSectionRequest'
      responses:
        '200':
          description: Section updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedSectionResponse'
...
components:
  schemas:
...
    RequirementsBlueprint:
      type: object
      properties:
        sections:
          type: array
          items:
            $ref: '#/components/schemas/SectionRequirement'
        eligibility_criteria:
          type: array
          items:
            type: string
          example: ["Serve population <5000", "Non-profit organization"]
        deadline:
          type: string
          format: date
          example: "2025-12-31"
        funding_amount:
          type: string
          example: "$50,000 - $100,000"
        scoring_criteria:
          type: object
          additionalProperties:
            type: integer
          example:
            need: 20
            approach: 30
            capacity: 25
            budget: 25
    SectionRequirement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Community Need Statement"
        order:
          type: integer
          example: 1
        word_limit:
          type: integer
          nullable: true
          example: 500
        char_limit:
          type: integer
          nullable: true
        format_requirements:
          type: string
          nullable: true
          example: "Narrative paragraph"
        scoring_weight:
          type: integer
          nullable: true
          example: 20
        generation_status:
          type: string
          enum: [not_started, generating, completed, failed]
          example: not_started
...
    GeneratedSectionResponse:
      type: object
      properties:
        section_id:
          type: string
          format: uuid
        section_name:
          type: string
          example: "Community Need Statement"
        final_text:
          type: string
          description: Merged AI text + user edits
        word_count:
          type: integer
          example: 487
        word_limit:
          type: integer
          nullable: true
          example: 500
        limit_exceeded:
          type: boolean
          example: false
        citations:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
        is_locked:
          type: boolean
          example: false
        generated_at:
          type: string
          format: date-time
        regeneration_count:
          type: integer
          example: 0
    Citation:
      type: object
      properties:
        citation_id:
          type: string
        position_in_text:
          type: integer
          description: Character offset in final_text
        document_id:
          type: string
          format: uuid
        document_title:
          type: string
          example: "Strategic Plan 2024"
        page_number:
          type: integer
          example: 5
        chunk_text:
          type: string
          description: Excerpt from source chunk
          example: "...our community lacks reliable internet..."
        relevance_score:
          type: number
          format: float
          example: 0.87
    UpdateSectionRequest:
      type: object
      properties:
        edited_text:
          type: string
          description: New text with user edits applied
      required: [edited_text]
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "File size exceeds 10MB limit"
        code:
          type: string
          example: "FILE_TOO_LARGE"
        details:
          type: object
          nullable: true